fix(rivetkit): route raw request fetches to actors
