+++ rivetkit-rust/engine/artifacts/errors/message.incoming_too_long.json
@@ -1,5 +1,5 @@
 {
   "code": "incoming_too_long",
   "group": "message",
-  "message": "Incoming message too long."
+  "message": "Incoming message too long"
 }
